requirements.txt
langchain
openai
streamlit
python-dotenv
langchain-community 
langchain-core
wikipedia
numexpr